# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project OMNI Fitness OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List exercises
      parameters:
        - in: query
          name: muscle
          schema:
            type: string
        - in: query
          name: equipment
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Exercise list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Create a custom exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExerciseBody"
      responses:
        "201":
          description: Created exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /exercises/{id}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get exercise details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
...
  /analytics/volume:
    get:
      operationId: getVolumeStats
      tags: [analytics]
      summary: Get volume stats over time
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            default: 30
        - in: query
          name: exerciseId
          schema:
            type: integer
      responses:
        "200":
          description: Volume stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VolumeDataPoint"
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Chat with AI coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
components:
  schemas:
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        primaryMuscle:
          type: string
        secondaryMuscles:
          type: array
          items:
            type: string
        equipment:
          type: string
          nullable: true
        trackingType:
          type: string
          enum: [weight_reps, distance_time, bodyweight_reps, time_only]
        instructions:
          type: string
          nullable: true
        isCustom:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - primaryMuscle
        - secondaryMuscles
        - trackingType
        - isCustom
        - createdAt
    CreateExerciseBody:
      type: object
      properties:
        name:
          type: string
        primaryMuscle:
          type: string
        secondaryMuscles:
          type: array
          items:
            type: string
        equipment:
          type: string
          nullable: true
        trackingType:
          type: string
          enum: [weight_reps, distance_time, bodyweight_reps, time_only]
        instructions:
          type: string
          nullable: true
      required:
        - name
        - primaryMuscle
        - trackingType
...
    VolumeDataPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        volume:
          type: number
        sets:
          type: integer
        reps:
          type: integer
        estimatedOneRepMax:
          type: number
          nullable: true
      required:
        - date
        - volume
        - sets
        - reps
...
    AiChatRequest:
      type: object
      properties:
        message:
          type: string
        apiKey:
          type: string
        context:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - message
        - apiKey
    AiChatResponse:
      type: object
      properties:
        response:
          type: string
        actionType:
          type: string
          enum: [message, create_workout, update_macros, generate_recipe]
          nullable: true
        actionData:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - response